I patched the HR timeseries query to cast value as numeric before averaging. This should fix the null results for HR in your weekly trends. Rebuild/redeploy and check if the HR chart now populates.

+++ app/routers/summary.py
@@ -304,7 +304,7 @@ async def space_series(request: Request, days: int = 30, conn = Depends(get_db))
                 agg as (
                   select
                     to_timestamp(floor(extract(epoch from start_time)/300.0)*300.0) as bucket_utc,
-                    avg(value) as hr
+                    avg(value::numeric) as hr
                   from gaia.samples
                   where user_id = %s
                     and type = 'heart_rate'
